@@ -3,26 +3,26 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@microsoft/signalr": "^5.0.8",
-    "babel-eslint": "^10.1.0",
-    "bootstrap": "^4.5.3",
+    "@babel/eslint-parser": "^7.19.1",
+    "@microsoft/signalr": "^6.0.10",
+    "bootstrap": "^5.2.2",
     "jquery": "^3.5.1",
     "merge": "^2.1.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-router": "^5.2.0",
-    "react-router-bootstrap": "^0.25.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "5.0.1",
-    "reactstrap": "^8.8.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router": "^6.4.2",
+    "react-router-bootstrap": "^0.26.2",
+    "react-router-dom": "^6.4.2",
+    "react-scripts": "^5.0.1",
+    "reactstrap": "^9.1.4",
     "rimraf": "^3.0.2"
   },
   "devDependencies": {
-    "ajv": "^7.0.3",
+    "ajv": "^8.11.0",
     "cross-env": "^7.0.3",
-    "eslint": "^7.17.0",
-    "eslint-config-react-app": "^6.0.0",
-    "eslint-plugin-flowtype": "^5.2.0",
+    "eslint": "^8.25.0",
+    "eslint-config-react-app": "^7.0.1",
+    "eslint-plugin-flowtype": "^8.0.3",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-react": "^7.22.0"
