@@ -1,11 +1,11 @@
 import React, { useState, useEffect } from "react";
-import { Route } from "react-router";
+import { Routes, Route } from "react-router";
 import { Layout } from "./components/Layout";
 import { Home } from "./components/Home";
 import { Lobby } from "./components/Lobby";
 import { Game } from "./components/Game";
 import { connect } from "./network";
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 
 import "./custom.css"
 
@@ -13,7 +13,7 @@ export default function App(props) {
     const [connection, setConnection] = useState(null);
     const [listOfTables, setListOfTables] = useState([]);
     const [players, setPlayerNames] = useState([]);
-    const history = useHistory();
+    const navigate = useNavigate();
 
     useEffect(() => {
         const newConnection = connect({
@@ -26,21 +26,23 @@ export default function App(props) {
 
     return (
       <Layout>
-        <Route exact path="/">
-            <Home
-                listOfTables={listOfTables}
-                selectTable={(table) => {
-                    connection.send("SelectTable", table);
-                    history.push(`/lobby/${table}`);
-                }}
-                saveTableName={(key, text) =>
-                    connection.send("SaveTableName", key, text)
-                } /></Route>
-        <Route path="/lobby">
-            <Lobby
-                playerNames={players}
-                savePlayerName={(key, text) => connection.send("SavePlayerName", key, text)}/></Route>
-        <Route path="/game"><Game /></Route>
+        <Routes>
+            <Route exact path="/" element={
+                <Home
+                    listOfTables={listOfTables}
+                    selectTable={(table) => {
+                        connection.send("SelectTable", table);
+                        navigate.push(`/lobby/${table}`);
+                    }}
+                    saveTableName={(key, text) =>
+                        connection.send("SaveTableName", key, text)
+                    } />} />
+            <Route path="/lobby" element={
+                <Lobby
+                    playerNames={players}
+                    savePlayerName={(key, text) => connection.send("SavePlayerName", key, text)}/>} />
+            <Route path="/game" element={<Game />} />
+        </Routes>
       </Layout>
     );
 }
